new tetrastream and filereader, but has bugs

+++ dm_density_plus/tetrahedron.h
@@ -9,18 +9,19 @@
 #ifndef TETRAHEDRON_H_
 #define TETRAHEDRON_H_
 #include <cmath>
+#include "types.h"
 
 class Point {
 public:
-	double x;
-	double y;
-	double z;
+	REAL x;
+	REAL y;
+	REAL z;
 };
 
 class Tetrahedron {
 public:
 	Tetrahedron();
-	double computeVolume();
+	REAL computeVolume();
 	Point v1;
 	Point v2;
 	Point v3;
@@ -28,18 +29,18 @@ public:
 
 	//test wheter the point is in this tetra
 	bool isInTetra(Point p);
-	double volume;
+	REAL volume;
 
-	double det4d(double m[4][4]);
-	void c2m(Point p1, Point p2, Point p3, Point p4, double m[4][4]);
+	REAL det4d(REAL m[4][4]);
+	void c2m(Point p1, Point p2, Point p3, Point p4, REAL m[4][4]);
 	Point getCenter();
 
-	double minx();
-	double miny();
-	double minz();
-	double maxx();
-	double maxy();
-	double maxz();
+	REAL minx();
+	REAL miny();
+	REAL minz();
+	REAL maxx();
+	REAL maxy();
+	REAL maxz();
 
 private:
 };
